FIX: erro era devido a ocorrencia de possivel multiplicação por 0 quando spectral_radius poderia ser 0

+++ src/core/Random.cpp
@@ -6,63 +6,53 @@
 #include <cmath>
 #include "nr_ran.h"
 
-Random::Random(const int function, const long seed)
-{
+Random::Random(const int function, const long seed) {
     idum = new long;
     *idum = seed;
 
     hasNextGaussian = false;
 
-    switch (function)
-    {
-    case RAN0:
-        ran = ran0;
-        break;
-    case RAN1:
-        *idum *= (seed > 0) ? -1 : 1; // guarantees idum < 0.
-        ran = ran1;
-        break;
-    case RAN2:
-        ran = ran2;
-    default: ;
+    switch (function) {
+        case RAN0:
+            ran = ran0;
+            break;
+        case RAN1:
+            *idum *= (seed > 0) ? -1 : 1; // guarantees idum < 0.
+            ran = ran1;
+            break;
+        case RAN2:
+            ran = ran2;
+        default: ;
     }
 }
 
-Random::~Random()
-{
+Random::~Random() {
     delete idum;
 }
 
-int Random::nextInt(const int max) const
-{
+int Random::nextInt(const int max) const {
     return static_cast<int>(floor(max * nextFloat()));
 }
 
-bool Random::nextBool() const
-{
+bool Random::nextBool() const {
     return nextFloat() < 0.5;
 }
 
-long double Random::nextGaussian(const long double m, const long double sd)
-{
-    long double next;
+double Random::nextGaussian(const double m, const double sd) {
+    double next;
 
     if (hasNextGaussian) // use value from previous call
     {
         next = nextNextGaussian;
         hasNextGaussian = false;
-    }
-    else
-    {
-        long double x1, x2, w;
+    } else {
+        double x1, x2, w;
 
-        do
-        {
+        do {
             x1 = 2.0 * nextFloat() - 1.0;
             x2 = 2.0 * nextFloat() - 1.0;
             w = x1 * x1 + x2 * x2;
-        }
-        while (w >= 1.0 || w == 0);
+        } while (w >= 1.0 || w == 0);
 
         w = sqrt((-2.0 * log(w)) / w);
         next = x1 * w;
@@ -73,17 +63,14 @@ long double Random::nextGaussian(const long double m, const long double sd)
     return (m + next * sd);
 }
 
-long double Random::nextFloat(const long double min, const long double max) const
-{
+double Random::nextFloat(const double min, const double max) const {
     return nextFloat() * (max - min) + min;
 }
 
-long double Random::nextFloat() const
-{
-    return (*ran)(idum);
+double Random::nextFloat() const {
+    return static_cast<double>((*ran)(idum));
 }
 
-long Random::getidum() const
-{
+long Random::getidum() const {
     return *idum;
 }
